Add Validation
Add validation for alreadycheckin
Add validation for null checkout
put useeffect into button
