+++ src/components/Content.css
@@ -10,16 +10,82 @@
     color: rgb(199, 17, 17);
 }
 
-.checkin_content {
-    margin-top: 40%;
-    margin-bottom: 30%;
-    text-align: center;
+@media screen and (min-width: 280px) {
+    .checkin_content {
+        margin-top: 10%;
+        margin-bottom: 20%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 10%;
+        margin-bottom: 20%;
+        text-align: center;
+    }
 }
 
-.checkout_content {
-    margin-top: 40%;
-    margin-bottom: 30%;
-    text-align: center;
+@media screen and (min-width: 320px) and (max-height : 600px){
+    .checkin_content {
+        margin-top: 10%;
+        margin-bottom: 10%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 10%;
+        margin-bottom: 10%;
+        text-align: center;
+    }
+}
+
+@media screen and (min-width: 320px) and (min-height : 700px){
+    .checkin_content {
+        margin-top: 20%;
+        margin-bottom: 30%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 20%;
+        margin-bottom: 30%;
+        text-align: center;
+    }
+}
+
+@media screen and (min-width: 400px) {
+    .checkin_content {
+        margin-top: 20%;
+        margin-bottom: 20%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 20%;
+        margin-bottom: 20%;
+        text-align: center;
+    }
+}
+
+@media screen and (min-width: 600px) {
+    .checkin_content {
+        margin-top: 30%;
+        margin-bottom: 30%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 30%;
+        margin-bottom: 30%;
+        text-align: center;
+    }
+}
+
+@media screen and (min-width: 1000px) {
+    .checkin_content {
+        margin-top: 5%;
+        margin-bottom: 5%;
+        text-align: center;
+    }
+    .checkout_content {
+        margin-top: 10%;
+        margin-bottom: 10%;
+        text-align: center;
+    }
 }
 
 .checkedin_msg {
@@ -41,28 +107,7 @@
     font-size: 20px;
 }
 
-@media (min-width:376px) {
-    .checkin_content {
-        margin-top: 5%;
-        text-align: center;
-        font-size: 25px;
-        line-height: 0.5;
-        margin-bottom: 5%;
-    }
-
-    .checkout_content {
-        margin-top: 5%;
-        text-align: center;
-        font-size: 25px;
-        line-height: 0.5;
-        margin-bottom: 16%;
-    }
-
-    .checkedin_msg {
-        margin-top: 100px;
-        margin-bottom: 100px;
-    }
-
+@media (min-width:1000px) {
     .traffic {
         display: flex; 
         justify-content: center;
@@ -73,6 +118,7 @@
         text-shadow: -1px -1px 1px rgba(255,255,255,.1), 1px 1px 1px rgba(0,0,0,.5);
     }
 }
+
